Add logger to Util
